made precision_id an integer and set precision.code as the pkey.

+++ db/migrate/20140325221028_create_precisions.rb
@@ -1,7 +1,7 @@
 class CreatePrecisions < ActiveRecord::Migration
   def change
     create_table :precisions do |t|
-      t.float :code
+      t.integer :code
       t.string :description
 
       t.timestamps
